added custom hooks for post functionalities to achieve SRP, and clean but centralized logics
